Move User update logic into User module
In this case, it doesn't add much.
I think this is because the responsibility of User is straightforward
for now: update the model when data are fetched.

Maybe we could move the "LogIn" logic into the User module also. Or not.

Anyway, this was preparatory work for the most interesting part: Tracks.
I expect to get more value by refactoring Tracks the same way.
Mostly: have a more appropriate encapsulation of Tracks logic and expose
less from this module.

Future will tell…